backend/extensions/daemon_proxy/requirements.txt
aiohttp>=3.8.0
docker>=6.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pyyaml>=6.0

